Fixes #11354 - Made the container deal with cpu shares and cpu sets correctly
According to docker remote api 'cpu shares' need to be integers.
However foreman docker sends it a float instead causing a
'cant convert to int64' error. This commit fixes that by ensuring what
gets sent to the container is an integer

Similarly According to docker remote api cpusets need to be strings.
However foreman docker sends it an integer instead causing a
'cant convert to string' error. This commit fixes that by ensuring what
gets sent to the container is an string